docs: restructure documentation and consolidate configuration
- Streamline README (EN/CN) as entry point with architecture overview
- Add docs/development.md covering dev workflow, sccache, and project architecture
- Translate docs/cluster.md to Chinese
- Simplify CONTRIBUTING.md and CLAUDE.md with cross-document pointers
- Remove dead docs: CHANGES.md, scripts/README.md
- Rename start_node_cluster.sh to cluster.sh for clarity
- Makefile delegates build targets to dev.sh, add standalone/cluster targets

Config changes:
- Remove dead config files: config.example.toml, cluster.conf, kiwi.conf
- Move RuntimeConfig into conf crate (no circular deps)
- Add --sample-config and --full-sample-config CLI flags
- Config::load() supports both TOML and Redis-style formats
- Config serialization to Redis-style via toml::Value flattening
