+++ AGENTS.md
@@ -1,0 +1,1 @@
+CLAUDE.md
