--- cluster.conf
@@ -1,43 +1,0 @@
-# Copyright (c) 2024-present, arana-db Community.  All rights reserved.
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# Kiwi Cluster Configuration Example
-# This file demonstrates how to configure Kiwi for cluster mode
-
-# Basic server settings
-port 7379
-binding 127.0.0.1
-
-# Cluster configuration
-cluster-enabled yes
-cluster-node-id 1
-cluster-members 1:127.0.0.1:7379,2:127.0.0.1:7380,3:127.0.0.1:7381
-cluster-data-dir ./raft_data
-cluster-heartbeat-interval 1000
-cluster-election-timeout-min 3000
-cluster-election-timeout-max 6000
-cluster-snapshot-threshold 1000
-cluster-max-payload-entries 100
-
-# Storage settings
-memory 1GB
-rocksdb-max-background-jobs 4
-rocksdb-max-write-buffer-number 2
-rocksdb-write-buffer-size 67108864
-
-# Logging
-log-dir /data/kiwi_rs/logs
