+++ src/common/runtime/Cargo.toml
@@ -18,6 +18,7 @@ num_cpus = { workspace = true }
 uuid = { workspace = true }
 resp = { workspace = true }
 storage = { workspace = true }
+conf = { path = "../../conf" }
 bytes = { workspace = true }
 snafu = { workspace = true }
 chrono = { version = "0.4", features = ["serde"] }
