+++ src/conf/Cargo.toml
@@ -9,10 +9,13 @@ workspace = true
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
 serde_ini = "0.2.0"
+toml = "0.8"
 validator = { version = "0.20", features = ["derive"] }
 snafu = "0.8.5"
 rocksdb.workspace = true
 openraft.workspace = true
+num_cpus = "1.16"
+log = "0.4"
 
 [dev-dependencies]
 tempfile = "3"
