+++ src/conf/src/config.rs
@@ -22,6 +22,7 @@ use validator::Validate;
 
 use crate::de_func::{parse_bool_from_string, parse_memory, parse_redis_config};
 use crate::error::Error;
+use crate::runtime_config::RuntimeConfig;
 
 /// Compression algorithm for RocksDB column families.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
@@ -83,7 +84,7 @@ impl std::str::FromStr for CompressionType {
 const DEFAULT_BINDING: &str = "127.0.0.1";
 const DEFAULT_PORT: u16 = 7379; // Redis-compatible port (7xxx variant of 6379)
 // config struct define - keeping original config items but using Redis-style format
-#[derive(Validate)]
+#[derive(Validate, Serialize, Deserialize)]
 pub struct Config {
     // Original config items from config.ini
     #[validate(range(min = 1024, max = 65535))]
@@ -121,6 +122,10 @@ pub struct Config {
     /// Authentication password. When set, clients must authenticate via AUTH command.
     pub requirepass: Option<String>,
     pub raft: Option<RaftClusterConfig>,
+
+    /// Dual-runtime configuration.
+    #[serde(default)]
+    pub runtime: RuntimeConfig,
 }
 
 impl std::fmt::Debug for Config {
@@ -197,11 +202,12 @@ impl std::fmt::Debug for Config {
                 },
             )
             .field("raft", &self.raft)
+            .field("runtime", &self.runtime)
             .finish()
     }
 }
 
-#[derive(Debug, Validate, Clone)]
+#[derive(Debug, Validate, Clone, Serialize, Deserialize)]
 pub struct RaftClusterConfig {
     #[validate(range(min = 1))]
     pub node_id: u64,
@@ -250,16 +256,25 @@ impl Default for Config {
             small_compaction_duration_threshold: 10000,
             requirepass: None,
             raft: None,
+            runtime: RuntimeConfig::default(),
         }
     }
 }
 impl Config {
-    // load config from file - supports Redis-style format with comments
+    // load config from file - supports TOML and Redis-style key-value format
     pub fn load(path: &str) -> Result<Self, Error> {
         let content =
             std::fs::read_to_string(path).context(crate::error::ConfigFileSnafu { path })?;
 
-        // Parse Redis-style configuration
+        // Try TOML first
+        if let Ok(config) = toml::from_str::<Config>(&content) {
+            config
+                .validate()
+                .map_err(|e| Error::ValidConfigFail { source: e })?;
+            return Ok(config);
+        }
+
+        // Fall back to Redis-style key-value format
         let config_map = parse_redis_config(&content).map_err(|e| Error::InvalidConfig {
             source: serde_ini::de::Error::Custom(e),
         })?;
@@ -577,6 +592,79 @@ impl Config {
             .map_err(|_e| Error::ValidConfigFail { source: _e })?;
 
         Ok(config)
+    }
+
+    /// Generate a sample configuration file from the default Config (Redis-style).
+    pub fn sample_config() -> String {
+        format!(
+            "# Kiwi Configuration File\n\
+             # Generated by `kiwi --sample-config`\n\n{}",
+            Config::default().to_redis_style()
+        )
+    }
+
+    /// Generate a complete sample configuration with all available keys, including
+    /// Raft cluster settings and authentication.
+    pub fn full_sample_config() -> String {
+        let mut c = Config::default();
+        c.raft = Some(RaftClusterConfig {
+            node_id: 1,
+            raft_addr: "127.0.0.1:8081".into(),
+            resp_addr: "127.0.0.1:7379".into(),
+            data_dir: "./raft_data".into(),
+            heartbeat_interval_ms: Some(200),
+            election_timeout_min_ms: Some(500),
+            election_timeout_max_ms: Some(1500),
+            use_memory_log_store: false,
+        });
+        c.requirepass = Some(String::new());
+
+        format!(
+            "# Kiwi Configuration File (all keys)\n\
+             # Generated by `kiwi --full-sample-config`\n\n{}",
+            c.to_redis_style()
+        )
+    }
+}
+
+impl Config {
+    /// Serialize to Redis-style `key value` format via `toml::Value` flattening.
+    /// Nested tables are flattened with `-` separators (e.g. `raft.node_id` → `raft-node-id`).
+    /// Field name underscores become hyphens.
+    fn to_redis_style(&self) -> String {
+        let value = toml::Value::try_from(self).expect("Config should serialize to toml::Value");
+        let mut lines: Vec<(String, String)> = Vec::new();
+        flatten_table(value.as_table().expect("top-level should be a table"), "", &mut lines);
+        lines.sort_by(|a, b| a.0.cmp(&b.0));
+        let mut out = String::new();
+        for (key, val) in lines {
+            out.push_str(&format!("{} {}\n", key, val));
+        }
+        out
+    }
+}
+
+fn flatten_table(table: &toml::Table, prefix: &str, lines: &mut Vec<(String, String)>) {
+    for (key, value) in table {
+        let flat_key = key.replace('_', "-");
+        let full_key = if prefix.is_empty() {
+            flat_key
+        } else if flat_key.starts_with(&format!("{}-", prefix)) {
+            // e.g. raft.raft_addr → raft-addr, not raft-raft-addr
+            flat_key
+        } else {
+            format!("{}-{}", prefix, flat_key)
+        };
+        match value {
+            toml::Value::Table(inner) => flatten_table(inner, &full_key, lines),
+            toml::Value::String(s) => lines.push((full_key, s.clone())),
+            toml::Value::Integer(i) => lines.push((full_key, i.to_string())),
+            toml::Value::Float(f) => lines.push((full_key, f.to_string())),
+            toml::Value::Boolean(b) => {
+                lines.push((full_key, if *b { "yes".into() } else { "no".into() }))
+            }
+            _ => {} // skip arrays, datetimes — not used in Config
+        }
     }
 }
